refactor: Move bounds code into its own file
